// src/app/interfaces/printing-house-article.ts
/**
 * Printing House Article Interface
 * @author Daniel Sogl
 */
export interface PrintingHouseArticle {
    /** Image heigh */
    heigh: number;
    /** Minimal price */
    minPrice: number;
    /** Image name */
    name: string;
    /** Costomer price */
    price: number;
    /** Image width */
    width: number;
  }